# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SlickByte API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * User-indicated portion size hint
 */
export type AnalyzeMealBodyPortion =
  (typeof AnalyzeMealBodyPortion)[keyof typeof AnalyzeMealBodyPortion];

export const AnalyzeMealBodyPortion = {
  small: "small",
  medium: "medium",
  large: "large",
} as const;

/**
 * Plate size used as a scale reference
 */
export type AnalyzeMealBodyPlateSize =
  (typeof AnalyzeMealBodyPlateSize)[keyof typeof AnalyzeMealBodyPlateSize];

export const AnalyzeMealBodyPlateSize = {
  small_plate: "small_plate",
  medium_plate: "medium_plate",
  large_plate: "large_plate",
} as const;

export interface AnalyzeMealBody {
  /** Base64-encoded image data (without data URL prefix) */
  imageBase64: string;
  /** Image MIME type (e.g. image/jpeg) */
  mimeType?: string;
  /** Optional extra context from user about the meal */
  prompt?: string;
  /** User-indicated portion size hint */
  portion?: AnalyzeMealBodyPortion;
  /** Plate size used as a scale reference */
  plateSize?: AnalyzeMealBodyPlateSize;
}

/**
 * AI confidence level
 */
export type MealAnalysisConfidence =
  (typeof MealAnalysisConfidence)[keyof typeof MealAnalysisConfidence];

export const MealAnalysisConfidence = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Macros {
  /** Protein in grams */
  protein: number;
  /** Carbohydrates in grams */
  carbs: number;
  /** Fats in grams */
  fats: number;
  /** Fiber in grams */
  fiber: number;
}

export interface MealAnalysis {
  /** Name of the detected food */
  foodName: string;
  /** Estimated calorie count */
  calories: number;
  /** AI confidence level */
  confidence: MealAnalysisConfidence;
  /** Quick nutritional insight (e.g. High carb, moderate protein) */
  quickInsight: string;
  macros: Macros;
  /** Health score out of 10 */
  healthScore: number;
  /** List of health suggestions */
  suggestions: string[];
  /** Detected ingredients */
  ingredients: string[];
  /** Whether the result came from a previous cached analysis */
  cached?: boolean;
  /** Optional note about cache reuse */
  note?: string;
}

/**
 * Type of insight to retrieve
 */
export type GetMealInsightBodyInsightType =
  (typeof GetMealInsightBodyInsightType)[keyof typeof GetMealInsightBodyInsightType];

export const GetMealInsightBodyInsightType = {
  protein_breakdown: "protein_breakdown",
  make_healthier: "make_healthier",
  fat_loss: "fat_loss",
} as const;

export interface GetMealInsightBody {
  mealAnalysis: MealAnalysis;
  /** Type of insight to retrieve */
  insightType: GetMealInsightBodyInsightType;
}

export interface MealInsight {
  /** Insight title */
  title: string;
  /** Detailed insight content */
  content: string;
  /** Actionable tips */
  tips: string[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}