# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SlickByte API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Upload a base64-encoded image and receive nutritional analysis
 * @summary Analyze a meal photo
 */
export const analyzeMealBodyMimeTypeDefault = `image/jpeg`;
export const analyzeMealBodyPortionDefault = `medium`;
export const analyzeMealBodyPlateSizeDefault = `medium_plate`;

export const AnalyzeMealBody = zod.object({
  imageBase64: zod
    .string()
    .describe("Base64-encoded image data (without data URL prefix)"),
  mimeType: zod
    .string()
    .default(analyzeMealBodyMimeTypeDefault)
    .describe("Image MIME type (e.g. image\/jpeg)"),
  prompt: zod
    .string()
    .optional()
    .describe("Optional extra context from user about the meal"),
  portion: zod
    .enum(["small", "medium", "large"])
    .default(analyzeMealBodyPortionDefault)
    .describe("User-indicated portion size hint"),
  plateSize: zod
    .enum(["small_plate", "medium_plate", "large_plate"])
    .default(analyzeMealBodyPlateSizeDefault)
    .describe("Plate size used as a scale reference"),
});

export const AnalyzeMealResponse = zod.object({
  foodName: zod.string().describe("Name of the detected food"),
  calories: zod.number().describe("Estimated calorie count"),
  confidence: zod
    .enum(["high", "medium", "low"])
    .describe("AI confidence level"),
  quickInsight: zod
    .string()
    .describe("Quick nutritional insight (e.g. High carb, moderate protein)"),
  macros: zod.object({
    protein: zod.number().describe("Protein in grams"),
    carbs: zod.number().describe("Carbohydrates in grams"),
    fats: zod.number().describe("Fats in grams"),
    fiber: zod.number().describe("Fiber in grams"),
  }),
  healthScore: zod.number().describe("Health score out of 10"),
  suggestions: zod.array(zod.string()).describe("List of health suggestions"),
  ingredients: zod.array(zod.string()).describe("Detected ingredients"),
  cached: zod
    .boolean()
    .optional()
    .describe("Whether the result came from a previous cached analysis"),
  note: zod.string().optional().describe("Optional note about cache reuse"),
});

/**
 * Get a smart follow-up insight about an analyzed meal
 * @summary Get a specific insight about a meal
 */
export const GetMealInsightBody = zod.object({
  mealAnalysis: zod.object({
    foodName: zod.string().describe("Name of the detected food"),
    calories: zod.number().describe("Estimated calorie count"),
    confidence: zod
      .enum(["high", "medium", "low"])
      .describe("AI confidence level"),
    quickInsight: zod
      .string()
      .describe("Quick nutritional insight (e.g. High carb, moderate protein)"),
    macros: zod.object({
      protein: zod.number().describe("Protein in grams"),
      carbs: zod.number().describe("Carbohydrates in grams"),
      fats: zod.number().describe("Fats in grams"),
      fiber: zod.number().describe("Fiber in grams"),
    }),
    healthScore: zod.number().describe("Health score out of 10"),
    suggestions: zod.array(zod.string()).describe("List of health suggestions"),
    ingredients: zod.array(zod.string()).describe("Detected ingredients"),
    cached: zod
      .boolean()
      .optional()
      .describe("Whether the result came from a previous cached analysis"),
    note: zod.string().optional().describe("Optional note about cache reuse"),
  }),
  insightType: zod
    .enum(["protein_breakdown", "make_healthier", "fat_loss"])
    .describe("Type of insight to retrieve"),
});

export const GetMealInsightResponse = zod.object({
  title: zod.string().describe("Insight title"),
  content: zod.string().describe("Detailed insight content"),
  tips: zod.array(zod.string()).describe("Actionable tips"),
});