# lib/api-zod/src/generated/types/analyzeMealBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SlickByte API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeMealBodyPlateSize } from "./analyzeMealBodyPlateSize";
import type { AnalyzeMealBodyPortion } from "./analyzeMealBodyPortion";

export interface AnalyzeMealBody {
  /** Base64-encoded image data (without data URL prefix) */
  imageBase64: string;
  /** Image MIME type (e.g. image/jpeg) */
  mimeType?: string;
  /** Optional extra context from user about the meal */
  prompt?: string;
  /** User-indicated portion size hint */
  portion?: AnalyzeMealBodyPortion;
  /** Plate size used as a scale reference */
  plateSize?: AnalyzeMealBodyPlateSize;
}