# lib/api-zod/src/generated/types/getMealInsightBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SlickByte API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GetMealInsightBodyInsightType } from "./getMealInsightBodyInsightType";
import type { MealAnalysis } from "./mealAnalysis";

export interface GetMealInsightBody {
  mealAnalysis: MealAnalysis;
  /** Type of insight to retrieve */
  insightType: GetMealInsightBodyInsightType;
}